Add fix Resistance Class

+++ src/PythagoreanTheorem.java
@@ -21,13 +21,13 @@ public class PythagoreanTheorem {
 
 		
 		// Get 'A' Value
-		Double a = Math.pow(m, 2) - Math.pow(n, 2);
+		double a = Math.pow(m, 2) - Math.pow(n, 2);
 
 		// Get 'B' Value
-		int b = 2 * m * n;
+		double b = 2 * m * n;
 		
 		// Get 'C' Value
-		Double c = Math.sqrt((Math.pow(m, 2) + Math.pow(n, 2)));
+		double c = Math.sqrt((Math.pow(m, 2) + Math.pow(n, 2)));
 				
 		
 		// Print the 'A' value
